move sidebar to its own route group
